post deletion problem fixed

+++ app/Http/Controllers/PostController.php
@@ -12,7 +12,7 @@ use Auth;
 class PostController extends Controller
 {
 	public function  getDashboard(){
-		$posts = Post::all();
+		$posts = Post::orderBy('created_at', 'desc')->get();
 		return view('dashboard')->with(['posts' => $posts]);
 	}
 	public function postCreatePost(Request $request)
@@ -31,6 +31,9 @@ class PostController extends Controller
 
 	public function deletePost($post_id){
 		$post = Post::where('id',$post_id)->first();
+		if (Auth::user() != $post->user) {
+			return redirect()->back();
+		}
 		$post->delete();
 		return redirect()->route('dashboard')->with('post_message', 'successfully deleted post');
 	}
